refactor(storage): extract the mcp_servers store out of db.rs into storage/mcp_store.rs
Pure MOVE, zero behavior change (the storage/links.rs pattern): the 5
mcp_servers CRUD methods (list/get/insert/delete_mcp_server,
set_mcp_server_consented) + migrate_mcp_servers + row_to_mcp_server move
verbatim into a sibling impl Db block; db.rs shrinks ~100 lines
(25,005 -> 24,905). storage/mod.rs registers the module; the methods stay
callable as db.method() (inherent impl, cross-file). Only migrate_mcp_servers
gains pub(crate) so migrate() can still call it; map_err/lock were already
pub(crate) from the links extraction.

No lock surface: mcp_servers rows are connection config (transport/args/
enabled/consented), not meeting content; the egress gating stays in the
connectors/mcp + tools tiers. Tests stay in db.rs (count 250 = trunk).
Verified: cargo test --lib mcp 39/0, clippy --lib -D warnings clean,
MSRV grep clean.
