+++ src-tauri/src/storage/mod.rs
@@ -1,5 +1,6 @@
 pub mod db;
 pub(crate) mod links;
+pub(crate) mod mcp_store;
 pub mod migration;
 pub mod models;
 pub mod usage;
